🍋 엔티티와 동일하게 만들어지도록 수정

+++ schema/community.sql
@@ -2,19 +2,18 @@ DROP TABLE community;
 
 
 CREATE TABLE community (
-	comNo INT NOT NULL AUTO_INCREMENT,
+	comNo LONG NOT NULL AUTO_INCREMENT,
 	comTitle VARCHAR(50) NOT NULL,
 	id VARCHAR(20) NOT NULL,
 	subject VARCHAR(15) NOT NULL,
 	comPw VARCHAR(6) ,
 	comRegdate DATETIME  DEFAULT now(),
 	comContent VARCHAR(255) NOT NULL,
-	comViewCount INT NOT NULL,
+	comViewCount LONG NOT NULL,
 	PRIMARY KEY(comNo)
 );
 
 ALTER TABLE community ADD constraint community_nick_fk FOREIGN KEY (id) REFERENCES studymembers(id);
-
